requirements.txt
Django==1.8
python-social-auth
dota2api